Add admin role to users and add authorization to products create action

+++ app/controllers/v1/products_controller.rb
@@ -24,37 +24,50 @@ class V1::ProductsController < ApplicationController
   end
 
   def update
-    product_id = params["id"]
-    product = Product.find_by(id: product_id)
-    product.name = params["input_name"] || product.name
-    product.price = params["input_price"] || product.price
-    # product.image_url = params["input_url"] || product.image_url
-    product.description = params["input_description"] || product.input_description
-    if product.save
-      render json: product.as_json
+    if current_user && current_user.admin 
+      product_id = params["id"]
+      product = Product.find_by(id: product_id)
+      product.name = params["input_name"] || product.name
+      product.price = params["input_price"] || product.price
+      # product.image_url = params["input_url"] || product.image_url
+      product.description = params["input_description"] || product.input_description
+      if product.save
+        render json: product.as_json
+      else
+        render json: {errors: product.errors.full_messages}, status: :unprocessable_entity
+      end
     else
-      render json: {errors: product.errors.full_messages}, status: :unprocessable_entity
+      render json: {}, status: :unauthorized
     end
   end
 
   def create
-    product = Product.new(
-      name: params["input_name"],
-      price: params["input_price"],
-      # image_url: params["input_url"],
-      description: params["input_description"]
-      )
-    if product.save
-      render json: product.as_json
+    if current_user && current_user.admin 
+      product = Product.new(
+        name: params["input_name"],
+        price: params["input_price"],
+        # image_url: params["input_url"],
+        description: params["input_description"],
+        supplier_id: 1
+        )
+      if product.save
+        render json: product.as_json
+      else
+        render json: {errors: product.errors.full_messages}, status: :unprocessable_entity
+      end
     else
-      render json: {errors: product.errors.full_messages}, status: :unprocessable_entity
+      render json: {}, status: :unauthorized
     end
   end
 
   def destroy
-    product_id = params["id"]
-    product = Product.find_by(id: product_id)
-    product.destroy
-    render json: {message: "Product successfully deleted!!!"}
+    if current_user && current_user.admin 
+      product_id = params["id"]
+      product = Product.find_by(id: product_id)
+      product.destroy
+      render json: {message: "Product successfully deleted!!!"}
+    else
+      render json: {}, status: :unauthorized
+    end
   end
 end
